Move Projects view from Feed to Stream; move FeedBanner to StreamBanner
