--- src/components/FeedBanner/index.js
@@ -1,3 +1,0 @@
-import component from './FeedBanner'
-
-export default component
