+++ src/components/StreamBanner/StreamBanner.scss
