+++ src/components/StreamBanner/index.js
@@ -1,0 +1,3 @@
+import component from './StreamBanner'
+
+export default component
