+++ src/components/TopicFeedHeader/TopicFeedHeader.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import Icon from 'components/Icon'
 import Button from 'components/Button'
-import { PostPrompt } from 'components/FeedBanner/FeedBanner'
+import { PostPrompt } from 'components/StreamBanner/StreamBanner'
 import { inflectedTotal, bgImageStyle } from 'util/index'
 import { DEFAULT_BANNER } from 'store/models/Group'
